third_party/blink/web_tests/fast/forms/select/select-disabled-expected.txt

Tests that user cannot change disabled select by keyboard

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS select1.value is "a"
PASS select1.value is "b"
PASS select1.value is "b"
PASS select1.value is "b"
PASS select2.value is "a"
PASS select2.value is "b"
PASS select2.value is "b"
PASS select2.value is "b"
Tests on select disable, popup menu is hidden
PASS internals.isSelectPopupVisible(select3) is true
PASS internals.isSelectPopupVisible(select3) is false
Tests on select disable, value cannot be changed
PASS select4.value is "b"
PASS select4.value is "b"
PASS successfullyParsed is true

TEST COMPLETE

a
b
c
 
a
b
c
 
a
 
a
b
c
 
